Agentic_RAG/requirements.txt
# Core LangChain packages
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.20

# Vector Store and Embeddings
faiss-cpu==1.8.0
sentence-transformers==3.0.1

# CrewAI
crewai==0.5.0

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2

# PDF Processing
pypdf==3.17.4

# Optional: Web scraping and search tools
beautifulsoup4==4.12.3
requests==2.32.3
google-search-results==2.4.2

# Optional: For enhanced functionality
numpy==1.26.4
pandas==2.2.2

# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.0.0
